inline GospelSlider into page.tsx, save SOF images locally
